♻️ Refactor: 카테고리 상수 공통화 및 스포츠 항목 제거
